pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-coldchain-guardian"
version = "0.1.0"
description = "Agentic AI template for Amazon Nova hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "pydantic==2.11.7",
  "boto3==1.39.11",
  "python-dotenv==1.1.1",
  "pyyaml==6.0.2"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
